Refactor: client 폴더 구조 리펙토링
* client 폴더 구조 리펙토링
