+++ client/src/App.js
@@ -1,24 +1,16 @@
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Main from "./pages/Main";
-import Signup from "./pages/Signup";
-import SellerSignup from "./pages/SellerSignup";
-import Login from "./pages/Login";
-import FindID from "./pages/FindID";
-import FindPW from "./pages/FindPW";
-import Product from "./pages/Product";
+import Member from "./pages/Member";
+import Find from "./pages/Find";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-          <Route path="/" element={<Main />} />
-          <Route path="/signup" element={<Signup />} />
-          <Route path="/seller/signup" element={<SellerSignup />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/findid" element={<FindID />} />
-          <Route path="/findpw" element={<FindPW />} />
-          <Route path="/product/:category" element={<Product />} />
+        <Route path="/" element={<Main />} />
+        <Route path="/member/:target" element={<Member />} />
+        <Route path="/find/:target" element={<Find />} />
       </Routes>
     </BrowserRouter>
   );
