+++ client/src/components/common/NavBar.js
@@ -1,4 +1,5 @@
 import "../../App.css";
+import "./common.css";
 import { Link } from "react-router-dom";
 
 function NavBar() {
