+++ client/src/components/common/SearchBar.js
@@ -1,4 +1,5 @@
 import "../../App.css";
+import "./common.css";
 import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCaretDown, faMagnifyingGlass, faCartShopping } from "@fortawesome/free-solid-svg-icons";
