+++ client/src/components/find/find.css
@@ -1,0 +1,59 @@
+/* --------------------
+  find css
+-------------------- */
+
+.find {
+  width: 360px;
+  margin: 100px auto;
+}
+
+.find > input {
+  width: 100%;
+  height: 50px;
+  font-size: 18px;
+  margin-top: 30px;
+  padding: 0 10px;
+}
+
+.find__title {
+  text-align: center;
+  font-size: 30px;
+  margin-bottom: 30px;
+}
+
+.find__auth-btn {
+  width: 50%;
+  height: 50px;
+  background-color: white;
+  border: none;
+  border-bottom: rgb(221, 221, 221) 1px solid;
+  cursor: pointer;
+}
+
+.find__auth-btn--select {
+  width: 50%;
+  height: 50px;
+  color: green;
+  background-color: white;
+  border: none;
+  border-bottom: green 1px solid;
+  cursor: pointer;
+}
+
+.find__err-msg {
+  text-align: center;
+  color: red;
+  margin-top: 10px;
+  margin-bottom: -20px;
+}
+
+.find__find-btn {
+  width: 100%;
+  height: 50px;
+  color: green;
+  background-color: white;
+  border: green 1px solid;
+  border-radius: 3px;
+  cursor: pointer;
+  margin-top: 30px;
+}
