+++ client/src/components/main/Product.js
@@ -1,4 +1,5 @@
-import "../App.css";
+import "../../App.css";
+import "./main.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faHeart } from "@fortawesome/free-solid-svg-icons";
