+++ client/src/components/member/SignupSeller.js
@@ -1,8 +1,9 @@
 import "../../App.css";
+import "./member.css";
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Check from "../../modal/Check";
+import Check from "../modal/Check";
 
 function SellerSignup() {
