+++ client/src/components/member/member.css
@@ -1,0 +1,165 @@
+/* --------------------
+  member css
+-------------------- */
+
+/* ---------- Login.js ---------- */
+
+.login {
+  width: 350px;
+  margin: 100px auto;
+}
+
+.login > input {
+  width: 100%;
+  height: 45px;
+  font-size: 18px;
+  margin-top: 20px;
+  padding: 0 10px;
+}
+
+.login > button {
+  width: 100%;
+  height: 50px;
+  color: green;
+  background-color: white;
+  border: green 1px solid;
+  border-radius: 3px;
+  cursor: pointer;
+  margin-top: 20px;
+}
+
+.login__title {
+  text-align: center;
+  font-size: 30px;
+  padding-bottom: 10px;
+}
+
+.login__err-msg {
+  text-align: center;
+  color: red;
+  margin-top: 10px;
+}
+
+.login__container {
+  display: flex;
+  margin-top: 10px;
+}
+
+.login__container__auto-login {
+  margin-right: auto;
+}
+
+.login__container__auto-login > label  {
+  display: block;
+  font-size: 14px;
+  cursor: pointer;
+}
+
+.login__container__auto-login input {
+  cursor: pointer;
+}
+
+.login__container__find-link {
+  display: flex;
+}
+
+.login__container__find-link > a {
+  font-size: 14px;
+}
+
+.login__container__find-link > hr {
+  height: 14px;
+  margin: auto 6px;
+}
+
+.login__signup {
+  text-align: center;
+  margin-top: 15px;
+}
+
+.login__signup svg {
+  color: gray;
+}
+
+/* ---------- Signup.js & SignupSeller.js ---------- */
+
+.signup {
+  width: 650px;
+  margin: 100px auto;
+}
+
+.signup span {
+  color: red;
+}
+
+.signup button {
+  width: 100%;
+  height: 100%;
+  color: green;
+  background-color: white;
+  border: green 1px solid;
+  border-radius: 3px;
+  cursor: pointer;
+}
+
+.signup__title {
+  text-align: center;
+  font-size: 30px;
+}
+
+.signup__require {
+  text-align: end;
+  color: gray;
+  border-bottom: black 2px solid;
+  font-size: 12px;
+}
+
+.signup__container {
+  display: flex;
+  margin: 30px 30px 0;
+}
+
+.signup__container__space-01 {
+  flex: 2.5;
+  margin: auto 0;
+}
+
+.signup__container__space-02 {
+  flex: 5.5;
+  height: 45px;
+}
+
+.signup__container__space-02 > input {
+  width: 100%;
+  height: 100%;
+  font-size: 18px;
+  padding: 0 10px;
+}
+
+.signup__container__space-02 svg {
+  margin-right: 6px;
+}
+
+.signup__container__space-03 {
+  flex: 2;
+  margin-left: 10px;
+}
+
+.signup__container__space-03 > button:disabled {
+  color: rgb(221, 221, 221);
+  border-color: rgb(221, 221, 221);
+  cursor: default;
+}
+
+.signup__err-msg {
+  text-align: center;
+  color: red;
+  margin-top: 10px;
+  margin-bottom: -20px;
+}
+
+.signup__btn {
+  width: 200px;
+  height: 50px;
+  margin: 30px auto 0;
+}
