+++ client/src/components/modal/modal.css
@@ -1,0 +1,40 @@
+/* --------------------
+  modal css
+-------------------- */
+
+/* ---------- Check.js ---------- */
+
+.check {
+  width: 350px;
+  height: 150px;
+  background-color: white;
+  margin: auto auto;
+}
+
+.check > button {
+  width: 100%;
+  height: 50px;
+  color: #7fad39;
+  background-color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.check__content {
+  display: flex;
+  height: 100px;
+  border-bottom: #7fad39 1px solid;
+  align-items: center;
+  justify-content: center;
+}
+
+.check-background {
+  display: flex;
+  background-color: rgba(0, 0, 0, 0.6);
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  z-index: 1;
+}
